requirements.txt
langchain
langchain-community
langchain-core
langchain_huggingface
langchain-text-splitters
langchain_groq
chromadb
langchain-chroma
pypdf
huggingface_hub
flask
python-dotenv
sentence-transformers
torch